Update ProfofConcept10000 (1).py
Added description of funcionality and new function dijkstra_to_target. To compute the shortest route from a start_vertex to a specific destination_vertex, you can define a modified version of the dijkstra() function. This version stops once the shortest path to the target is found — making it more efficient when you only need one route